% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-col-names.R
\name{clean_col_names}
\alias{clean_col_names}
\title{Clean up data frame column names}
\usage{
clean_col_names(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A data frame.
}
\description{
Clean up data frame column names
}
