% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Perform clustering}
\usage{
cluster(
  x,
  method = c("leiden"),
  resolution = 1,
  n_neighbors = 50,
  out_dir = NULL
)
}
\arguments{
\item{x}{A \linkS4class{SpatialExperiment} object.}

\item{method}{Clustering method (only `leiden` is implemented currently).}

\item{resolution}{Value of the parameter controlling the coarseness of the clusters (higher resolution yields more clusters).}

\item{n_neighbors}{The maximum number of nearest neighbors to compute.}

\item{out_dir}{Name of the output analysis directory. If specified, the object and the corresponding plots will be saved there.}
}
\value{
A SpatialExperiment object.
}
\description{
Cluster similar cells based on their expression profiles.
}
