% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-object.R
\name{create_object}
\alias{create_object}
\title{Create a SpatialExperiment object}
\usage{
create_object(
  x,
  expression_cols = NULL,
  metadata_cols = NULL,
  skip_cols = NULL,
  clean_names = TRUE,
  transformation = NULL,
  out_dir = NULL
)
}
\arguments{
\item{x}{A data frame or a path to a file (can be gzipped).}

\item{expression_cols}{Column names of markers/antibody expression as a vector or a grep pattern. Auto-detected if not specified.}

\item{metadata_cols}{Column names of cell metadata (not markers/antibodies) as a vector or a grep pattern. Auto-detected if not specified.}

\item{skip_cols}{Column names to ignore (such as bad antibodies) as a vector or a grep pattern.}

\item{clean_names}{A logical scalar. Clean the data frame column names to remove problematic characters and make them unique.}

\item{transformation}{A character string indicating which transformation method should be used. See \code{\link[=transform]{transform()}}.}

\item{out_dir}{Name of the output analysis directory. If specified, the object and the corresponding plots will be saved there.}
}
\value{
A \linkS4class{SpatialExperiment} object.
}
\description{
Create a \linkS4class{SpatialExperiment} object (used in the same manner as \linkS4class{SingleCellExperiment}) that will store all of the project data.
}
