% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot expression distribution}
\usage{
plot_distribution(x, assay = "exprs", out_dir = NULL)
}
\arguments{
\item{x}{A \linkS4class{SpatialExperiment} object.}

\item{assay}{A character string indicating which values should be used.}

\item{out_dir}{Name of the output analysis directory. If specified, the plots will be saved there.}
}
\value{
.
}
\description{
Plot distribution of expression intensities, split by marker.
}
