% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dr.R
\name{plot_dr}
\alias{plot_dr}
\title{Plot dimensionality reduction}
\usage{
plot_dr(
  x,
  dr,
  color_by,
  assay = "logcounts",
  smooth = FALSE,
  range = c(0.01, 0.99),
  out_dir = NULL
)
}
\arguments{
\item{x}{A \linkS4class{SpatialExperiment} object.}

\item{dr}{.}

\item{color_by}{Column metadata field(s) or feature(s) to color by.}

\item{assay}{A character string indicating which expression values should be used.}

\item{smooth}{A logical scalar. Smooth values. Helps to visualize expression patterns in a plot with many overlapping points.}

\item{range}{A vector of 2 values indicating the minimum and maximum percentiles for the color range. Helps to visualize expression patterns when extreme outliers are present. For example, \code{c(0, 0.99)} will not expand the color scale above 99th percentile.}

\item{out_dir}{Name of the output analysis directory. If specified, the plots will be saved there.}
}
\value{
.
}
\description{
Plot cell-level reduced dimension results stored in a SpatialExperiment object.
}
