% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Generate a heatmap}
\usage{
plot_heatmap(x, group_by, assay = "logcounts", out_dir = NULL)
}
\arguments{
\item{x}{A \linkS4class{SpatialExperiment} object.}

\item{group_by}{Column metadata field(s) to group by.}

\item{assay}{A character string indicating which expression values should be used.}

\item{out_dir}{Name of the output analysis directory. If specified, the plots will be saved there.}
}
\value{
A \linkS4class{Heatmap} object.
}
\description{
Generate a heatmap
}
