% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Generate a generic scatter plot}
\usage{
plot_scatter(
  data,
  x,
  y,
  color_by,
  smooth = FALSE,
  range = c(0.01, 0.99),
  title = "",
  aspect_ratio = 1
)
}
\arguments{
\item{data}{A data frame.}

\item{x}{.}

\item{y}{.}

\item{color_by}{Column metadata field(s) or feature(s) to color by.}

\item{smooth}{A logical scalar. Smooth values. Helps to visualize expression patterns in a plot with many overlapping points.}

\item{range}{A vector of 2 values indicating the minimum and maximum percentiles for the color range. Helps to visualize expression patterns when extreme outliers are present. For example, \code{c(0, 0.99)} will not expand the color scale above 99th percentile.}

\item{title}{Plot title.}

\item{aspect_ratio}{Aspect ratio of the panel.}
}
\value{
A ggplot object.
}
\description{
Create a scatter plot to display the relationship between two continuous variables.
This is essentially a wrapper for \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.
}
