% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform/normalize expression data}
\usage{
transform(x, method = c("log", "z"), out_dir = NULL)
}
\arguments{
\item{x}{A \linkS4class{SpatialExperiment} object.}

\item{method}{A character string indicating which transformation method should be used.}

\item{out_dir}{Name of the output directory. If specified, the density plots will be saved there.}
}
\value{
A \linkS4class{SpatialExperiment} object with transformed values in the \code{exprs} assay.
}
\description{
Transform/normalize expression data.
The options are currently:
\itemize{
  \item \code{log}: log10 + 1
  \item \code{z}: Z normalization of each marker (recommended by Hickey et al.)
}
}
