% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-exprs-cols.R
\name{detect_exprs_cols}
\alias{detect_exprs_cols}
\title{Detect marker/antibody column names}
\usage{
detect_exprs_cols(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A vector of column names.
}
\description{
Detect marker/antibody column names
}
